starting refactor, episodeTableViewCell successfully refactored
