+++ Podcast/SeriesFeedElementSubjectView.swift
@@ -1,0 +1,20 @@
+//
+//  SeriesFeedElementSubjectView.swift
+//  Podcast
+//
+//  Created by Natasha Armbrust on 10/7/17.
+//  Copyright © 2017 Cornell App Development. All rights reserved.
+//
+
+import UIKit
+
+class SeriesSubjectView: UIView {
+    
+    init() {
+        super.init(frame: CGRect.zero)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
