+++ Podcast/SeriesFeedElementSupplierView.swift
@@ -1,0 +1,19 @@
+//
+//  SeriesFeedElementSupplierView.swift
+//  Podcast
+//
+//  Created by Natasha Armbrust on 10/7/17.
+//  Copyright © 2017 Cornell App Development. All rights reserved.
+//
+
+import UIKit
+
+class SeriesFeedElementSupplierView: UIView {
+    init() {
+        super.init(frame: CGRect.zero)
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
